requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pydantic
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
pypdf
faiss-cpu